LayoutTests/fast/css/background-clip-text-expected.txt

This tests checks that the '-webkit-text' and 'text' keywords are parsed correctly in the 'background-clip' and '-webkit-background-clip' properties, and that 'background-clip' is parsed correctly in the 'background' shorthand.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("background-clip: -webkit-text", "background-clip") is "-webkit-text"
PASS test("background-clip: -webkit-text", "-webkit-background-clip") is ""
PASS test("background-clip: content-box", "background-clip") is "content-box"
PASS test("background-clip: padding-box", "background-clip") is "padding-box"
PASS test("background-clip: border-box", "background-clip") is "border-box"
PASS test("background-clip: text", "background-clip") is "text"
PASS test("background-clip: text", "-webkit-background-clip") is ""
PASS test("-webkit-background-clip: -webkit-text", "background-clip") is ""
PASS test("-webkit-background-clip: -webkit-text", "-webkit-background-clip") is "-webkit-text"
PASS test("-webkit-background-clip: text", "background-clip") is ""
PASS test("-webkit-background-clip: text", "-webkit-background-clip") is "text"
PASS test("background: url() padding-box", "-webkit-background-clip") is ""
PASS test("-webkit-mask: url() ", "background-clip") is ""
FAIL test("background: url() ", "background-clip") should be border-box. Was initial.
PASS test("background: url() padding-box", "background-clip") is "padding-box"
PASS test("background: url() text", "background-clip") is "text"
PASS test("background: url() padding-box border-box", "background-clip") is "border-box"
PASS test("background: url() padding-box text", "background-clip") is "text"
PASS test("background: repeat content-box border-box url() ", "background-clip") is "border-box"
PASS test("background: repeat content-box text url() ", "background-clip") is "text"
PASS test("background: padding-box blue content-box url() repeat scroll ", "background-clip") is "content-box"
PASS test("background: padding-box blue text url() repeat scroll ", "background-clip") is "text"
PASS successfullyParsed is true

TEST COMPLETE
